fix(topographer): own activation report facts (#846)
## Context

Part of the Coherence Cleanup stack. This branch moves activation report derivation to its owner package, @ontrails/topographer.

## What Changed

- Added Topographer activation report owner API and tests.
- Exported deriveActivationGraph, deriveDeclaredTrailActivation, deriveSignalActivationRelations, and related report types from @ontrails/topographer.
- Reduced the Trails app activation module to a compatibility re-export.
- Documented the activation report exports in the Topographer README and central API reference.
- Added branch-local changesets for @ontrails/topographer and @ontrails/trails.

## Verification

- bun test packages/topographer/src/__tests__/activation-report.test.ts && bun run --cwd packages/topographer typecheck
- bun run --cwd apps/trails typecheck && bun test apps/trails/src/__tests__/survey.test.ts apps/trails/src/__tests__/guide.test.ts
- bun run --cwd packages/topographer lint && bun run --cwd apps/trails lint
- bun run docs:wrap-check && bun run docs:links && bun run changeset:check && git diff --check
- bun run --cwd packages/topographer test

## Review

- In-thread branch local review was clean with no P0/P1/P2 findings.
- In-thread full-stack review found one relevant P3 docs gap for the new Topographer activation exports; fixed in this branch.
- Reviewers verified owner API export, app compatibility re-export, dependency availability, and changeset.
