+++ .changeset/topographer-activation-report-facts.md
@@ -1,0 +1,7 @@
+---
+"@ontrails/topographer": patch
+"@ontrails/trails": patch
+---
+
+Move activation report derivation into Topographer and keep the Trails app
+consuming the owner-held activation facts through a compatibility re-export.
